SafeTI library independization
All components intrinsic to the injector have been revised to drop the dependency of the grlib library (types, functions and component instation). However, now there must be a wrapper AHB interface for the target platform, as the 'injector_ahb_SELENE.vhd' for the Selene platform, which requires libraries from both the BSC and target platform.