Dependency on external libraries has been removed
Types, functions and internal connections have been remade to drop the dependency with Gaisler libraries. The project passes Spyglass, Vivado compilations and Questa simulation. However, now there is two top levels, the AHB interface (that contains nothing, because the AHB interface used is from an external library) and an SELENE/platform interface to translate from our types and connections to the platform's.