Small width change at BM bus from read/write_if
Since the size bus of the BM buses from read_if_bmo and write_if_bmo are already -1 in value respect the internal curr_size, they have been modified to match busses width of bm_out, which sends to the AHB interface the control information to make the transfer. This is a simpler modification to reduce complexity and does not represent any change at injecting performance.