Testbench half operative
The debugging of the testbench has allowed to learn better how it communicates with the interface over the BM bus. This information has been documented at 'Transaction_manual.txt' on doc's folder. It still requires a bit of work, and also tweak the injector so the RST control flag also resets the FIFO, if the team thinks that it should do it.