Testbench improved
It has been found that the testbench may hang up waiting for a injector request or testbench grant request assertion on specific cases. This has been solved with the improvement. Now, the user can set a threshold of how many clock cycles the test can wait to the requests and grant requests signals to be asserted, and outputing an error if the test pass that threshold of waiting.

Furthermore, the next improvement should be to add random waits to prolongue the injector/testbench waiting for a grant request/request, tacking into account the threshold set. So, the higher is the threshold, the more clock cycles may make to wait the injector/testbench in a random fashion. This could be used at future injector versions, that may reset transactions if the interface is not responding.