Vivado's AXI4 subordinate does not support narrow transactions
After the whole day trying to make the AXI4 subordinate from Vivado IP generator tool work on transfers of a size narrower than the AXI data bus width, I came to the conclusion that it just doesn't support narrow size transactions. Even found on a specification sheet (that may not be for the same thing, there're many called AXI4-something) that indicates that it is not recommended to make narrow transactions. Gonna try a bit with another subordinate from elsewhere or at the end, just try to debug it as well as I can with the Vivado's one.