Restructuration of the interface to make compatible subordinates that do not support narrow transaction + AXI read buffer for higher performance
Finding out that the AXI subordinate from Vivado IP library doesn't allow narrow transactions took me thinking that maybe, even though it doesn't follow the standard, it might be a design choice on other designs, since using the maximum of the AXI data bus is the most performance choice. Thus, I decided to re-implement the decide_size function and other things to support subordinates that do not support narrow transfers, even though the manager still does.

Furthermore, theimplementation of separation registers between manager and BM component/AXI interconnect or AXI subordinate (to maximize frequency of operation by increasing latency by 2 cc) has shown a mismatch of too many clock cycles, to the point that the latency would require to increase linearly with the number of beats of the transaction. Because this is unrealistic of how bad could be, decided to restructure the manager to execute AXI reads over a mini-buffer FIFO that the BM transfer logic will take to transfer to the BM component. This way, the AXI burst is done leaving one clock cycle between beats until the buffer is full, since it's expected that the BM transfer will be the bottleneck on most cases.