SafeTI descriptor program reset through software + AXI4 Manager v0.8.6
The program that the SafeTI injector loads now can be erased through software by setting to high the reset bit on the APB control register of the module.

The AXI4 Manager v0.8.6 has dropped the utility of 'Injector_implementation' due to lack of functionality. Additionally, it has been added input ports to be controlled by the BM component (SafeTI) to set specific AXI CACHE and PROT configuration on the AXI bursts, plus the signal to do fixed address access. However, at the moment these are hardcoded on the wrappers, since the SafeTI still does not support these capabilities.