SafeTI APB serial to load the injector program
Due to difficulties for simulating the injector on the SELENE platform, it has been decided to change the programming algorithm from "loading from main memory" to "APB serial", the latter being that the injector now is programmed by writing all the descriptors on the same APB register directly.

This commit also includes major changes on the descriptor layout, simplifying the descriptor from 5 words to 2. Thou this will not be the final form of the descriptor.

Furthermore, this means that many parts of the documentation are now outdated, but no new documentation will be produced until a more definitive version of the descriptor is made.