Rework on READ_IF and WRITE_IF modules + CI missname
Due to problematic behaviour, it has been decided to reformat the
modules for executing read and write transactions, also including
the testbench setting.

In addition, it has been found that the CI lint job would check
on "injector" folder instead of "injector_axi". This meant that,
even thou no folder was found, the tests would pass without an
error. Thus, a synthesis error has been hiding until now on the
AXI interface. This will be solved in the future.