AXI4 Manager interface v0.8.9 debugged
However, only injector_mode = TRUE is implemented and tested.
This mode generates the correct signaling to carry out the
transaction as a normal interface, but discards all read data and
zeroes all write data in order to finish as earlier as possible
the transaction at the BM side, even if it's still ongoing on the
AXI network.