Some tweaks to allow SafeTI synthesis + Documentation WIP
Some teaks have been done to ensure synthesis on cases where the data buses
between injector and interface are of different widths.

Furthermore, new documentation is being generated at the moment.
This will include a manual to aid on the design of injection software using or expanding
the available drivers at the folder bsc_safeti/sw, implementing the SafeTI on other
platforms, developing new interfaces and even modifying the core components of the injector
itself.