Readded support for the AHB SELENE interface + WRITE freeze bug correction
This commit modifies the AHB wrapper that includes the AHB Master interface
from SELENE's platform, to make it compatible with the changes made on the
SafeTI core.
However, READ_FIX and WRITE_FIX descriptors do not work correctly on the
AHB interface due to lack of native support on the protocol.
A small submodule could be implemented on the AHB wrapper to enable
compatibility wih the feature, but since the same traffic can be
obtained with READ and WRITE descriptors, this is put on standby.

The re-implementation of the AHB interface has shown a bug where the WRITE
and WRITE_FIX descriptors would halt from progresing if the interface does
not comply with the correct timing on the data reads pulses.
This has been solved and the injector core should now just set an error
flag, but otherwise, continue with the descriptor execution as normal
(which is what happens with the AHB interface at the moment).