Unintended fetch error solved on Out of Memory cases
It has been found that the FETCH stage would report an error when-
ever the write pointer would overflow the memory or when the last
descriptor on memory would be read.

These problems have been solved to only report the error if an
actual write is happening out of bounds (hence, out of memory) and
now the last descriptor is checked for 'last' asserted to report or
not the error on read (aka. if it's the last, continue operation,
if the last descriptor on the whole memory is not the last, then
report error to control).