SEQ descriptors added + SW updated+ APB generics SELENE AHB wrapper
The READ_SEQ and WRITE_SEQ descriptors have been developed and
added to the SafeTI core. These operations are able of accessing
sequentially on memory, just like normal READ and WRITE descriptors,
but it fixes a small access size for every access repetition.
Hence, the 'count' field is used for indicating how many accesses
of 'size' must be done sequentially (incrementing address) on memory.

Note that the descriptor word 0, the control word, is different on
SEQ descriptors than previous descriptors. The changes are done for
allowing a higher sequential access rather than be able to program
high sizes. This is particularly useful if the goal is to mimic
another component access, eg. CPU.

However, these have been partially debugged on platform simulation.
Future testbench should be made for further testing.

The drivers have been updated, which leaves the software documentation
a bit obsolete. This should be also updated in the future.