Drivers update to support multiple SafeTI modules
The drivers have been updated to support multiple SafeTI instances.
The only necessary changes to make it work is to match the base
addresses set on injector.h to where a SafeTI module is allocated
and use the number of the module on the software to communicate
with that particular SafeTI module.