Counters + MAX_SIZE_BURST+1 bug solved + AHB interface compatibility
The APB interface now includes two counter registers, one for
counting the number of interruptions sent, and other for
counting the number of transactions requests made to the
traffic interface (of course, with a maximum size of
MAX_SIZE_BURST).

In addition, a bug has been found where, on transactions of +1 bytes
over the maximum allowed per burst, the second transfer of one single
byte would not happen.

Furthermore, now the request grant is considered also as sending
one beat of transfer on writes. This has been done to solve a
compatibility problem with the Selene's AHB interface.
This means that the testbenches must be revised to include this
different behavior in the future.