Testbench update + Driver/HW corrections
The testbench package has been revised to support the last change,
the write request granting counts as a write beat. This however
supposes that the AXI interface now will signal size count  errors,
but otherwise it should work just fine.

The drivers and HW have been updated to include the lastest calls
for interacting with the new counters.