Fix Delay counter for SCA prot.+Driver uintptr format
The counters used by the DELAY descriptor weren't meant to be counting when the traffic injector
is disabled. This has been resolved, being necessary for the study on SCA Hybrid protection.

Furthermore, the drivers now use uintptr format for the generating the final address access to
the traffic injector APB interface.